Рефакторинг репозитория

+++ backend/src/main/java/ru/innopolis/stc27/maslakov/enterprise/project42/entities/food/Food.java
@@ -1,5 +1,6 @@
 package ru.innopolis.stc27.maslakov.enterprise.project42.entities.food;
 
+
 import lombok.Data;
 
 @Data
@@ -7,6 +8,6 @@ public class Food {
 
     private final int id;
     private String name;
-    private int price;
+    private double price;
     private FoodCategory foodCategory;
 }
